+++ tests/test_jira.py
@@ -4,7 +4,8 @@ from unittest import mock, TestCase
 from dateutil.tz import datetime
 from dateutil.tz.tz import tzutc
 
-from bugwarrior.config import ServiceConfig, BugwarriorConfigParser
+from bugwarrior.config import ServiceConfig
+from bugwarrior.config.load import BugwarriorConfigParser
 from bugwarrior.services.jira import JiraService
 
 from .base import ServiceTest, AbstractServiceTest
